TASK-6: Filter job + logging

+++ src/test/java/home/yura/websearchgui/TestUtils.java
@@ -90,7 +90,7 @@ public class TestUtils {
                 randomString(),
                 resultEntryDefinitionId,
                 filterItemId,
-                randomLong(),
+                (long) randomInt(),
                 randomString(),
                 false);
     }
